Enemy miniaturization Air Bubbles and Wave
Today I replaced the waves in all scenes with the prefab I made.  It works well

I also miniaturzed the water and flyinng enemies. They are a quarter scale, and their variables are adjusted to a quarter of previous values.

The enemies need to have a state for when they bang into walls. I need them to reverse direction. Currently they just stop. Work on this next time.

I also created the idea of air bubbles ( Just the heart shape for now). Basically when you collide with the heart shape, the air supply gets 10 added to it. I should keep working on that.

I'm also considering making an endless level and a few more mission levels.